refactor: split use case interfaces into separate files
Follow project rule: "One public class/interface per file"
