+++ common-sdk/src/main/kotlin/com/algorand/wallet/jointaccount/transaction/domain/usecase/AddJointAccountSignature.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2022-2025 Pera Wallet, LDA
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License
+ */
+
+package com.algorand.wallet.jointaccount.transaction.domain.usecase
+
+import com.algorand.wallet.foundation.PeraResult
+import com.algorand.wallet.jointaccount.transaction.domain.model.AddSignatureInput
+import com.algorand.wallet.jointaccount.transaction.domain.model.JointSignRequest
+
+fun interface AddJointAccountSignature {
+    suspend operator fun invoke(
+        signRequestId: String,
+        addSignatureInput: AddSignatureInput
+    ): PeraResult<JointSignRequest>
+}
